Support for a new field - slabMD for slabs.
new metadata field added to slabs. The additional data hopefully makes life simpler
for the parser.

+++ quarry.h
@@ -11,7 +11,7 @@ The name of the Hemanth Kapila may NOT be used to endorse or promote products de
 */
 #ifndef __QUARRY__
 #define __QUARRY__
-
+#include <sys/types.h>
 
 enum quarry_SlabType{
   quarry_Error,
@@ -37,7 +37,7 @@ struct quarry_Slab_
   int slabLength;
   enum quarry_SlabType slabType;
   unsigned char *data;
-  int slabMD;
+  u_int64_t slabMD;
 };
 
 typedef struct quarry_Slab_ quarry_Slab;
